Added the Customer Activity Detail API
The Customer Order summary would be used to obtain a summary listing of
all customer activity.  This request will be used when a detail listing
of a specific activity is request. Normally this request would follow
the summary request after the customer selected which activity more
detail is desired.  The activity number passed in can be an Order
Number or a Return Number.