Re-Factored the way sigma request fields are initialized
Instead of including a constant in each api class set as an array of
the required sigma request fields, I am now loading those fields
directly from the yam file associated with each api.  Because of this I
was able to move all initialization requirements to the TransactionSet
class.  You will now only need to add an initialize method to
individual apis if additional functionality is required, i.e. the
ProductAvailability class.