Re-factored how the request fields are sorted before sending to sigma
Instead of sorting the request fields right before they are sent to sigma, I opted to just put the the request field into the right index of the array from the start.  The main reason for this is because of 200+ request fields for orders. Before, if you needed field number 199 you would have to define all fields before that in the yaml file.  Now, you would no longer have to do that, you can simply define only what you absolutely need in the yaml file.